djando_project/requirements.txt
asn1crypto==0.24.0
astroid==2.0.4
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
colorama==0.4.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.4.2
Django==2.0
django-crispy-forms==1.7.2
django-jinja==2.4.1
django-jinja-bootstrap-form==4.2.3
django-mysql==2.4.1
django-rest-coreapi-schema==0.1.2
django-rest-framework==0.1.0
djangorestframework==3.7.7
geoip2==2.9.0
gunicorn==19.9.0
idna==2.7
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
jsonserver==0.1.0
lazy-object-proxy==1.3.1
Markdown==2.6.10
MarkupSafe==1.1.0
maxminddb==1.4.1
mccabe==0.6.1
mysql-connector-python==8.0.11
mysqlclient==1.3.10
Pillow==5.3.0
protobuf==3.6.1
psycopg2-binary==2.7.6.1
PyAudio==0.2.11
pycparser==2.19
Pygments==2.2.0
pylint==2.1.1
PyMySQL==0.9.2
pytz==2018.7
requests==2.21.0
simplejson==3.16.0
six==1.11.0
typed-ast==1.1.0
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==16.1.0
wrapt==1.10.11
